+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `CompletionCandidates.Fixed` now has a secondary convenience constructor that take a `vararg` of `String`s
 - `CompletionCadidates.Custom`, which allows you to call other binaries or write a script to generate completions.
+- `Option.wrapValue` and `Argument.wrapValue` to make it easier to reuse existing conversion functions.
 
 ### Changed
 - `option()` and `argument()` now take optional `completionCandidates` parameters to override how completion is generated. The constructor and `copy` functions of `OptionsWithValues` and `ProcessedArgument` have changed to support default values.